feature: (Create)
Create a check_files.py
