+++ score.txt
@@ -1,3 +1,4 @@
 GRESSIER;999
 GUILLON;777
 Alain;69
+Alain;69
